refactor: move strategies under their own folder
